Control.Error.Class

Description

This module exports error- and failure-handling combinators that are defined polymorphically. This means that when you use these combinators you don't commit to a specific monad transformer stack in your return type.

Throwing errors and failing

throwError :: MonadError e m => forall a. e -> m a

Is used within a monadic computation to begin exception processing.

liftEither :: MonadError e m => Either e a -> m a

Lift an Either value into any monadic computation by calling throwError on Left values and returning Right values.

mapWithError :: MonadError f m => (e -> f) -> (a -> b) -> ExceptT e m a -> m b

Map over a computation that can throw errors, rethrowing errors by applying a transformation.

mzero :: MonadPlus m => forall a. m a

the identity of mplus. It should also satisfy the equations

mzero >>= f  =  mzero
v >> mzero   =  mzero

liftMaybe :: MonadPlus m => Maybe a -> m a

Lift a Maybe into any MonadPlus.

Converting between MonadPlus and MonadError

note :: MonadError e m => e -> Maybe a -> m a

Tag the Nothing value of a Maybe.

See ?? for an infix version of note.

(??) :: MonadError e m => Maybe a -> e -> m a

Lift a Maybe value into any monad by calling throwError if it is Nothing.

This is the infix definition of note.

noteM :: MonadError e m => e -> MaybeT m a -> m a

Rewrite a MonadPlus constraint into a MonadError constraint by throwing an error if the computation fails with mzero.

See !? for an infix version of noteM.

(!?) :: MonadError e m => MaybeT m a -> e -> m a

Handle failure in a computation by throwing an error if it fails.

This is the infix definition of noteM.

hushM :: MonadPlus m => ExceptT e m a -> m a

Rewrite a MonadError constraint into a failure under MonadPlus by rewriting throwError _ as mzero

Eliminating MonadError constraints

exceptT :: Monad m => (e -> m b) -> (a -> m b) -> ExceptT e m a -> m b

Fold an ExceptT by providing one continuation for each constructor.

This can be used to eliminate a MonadError constraint from a computation.

catchError :: Monad m => ExceptT e m a -> (e -> m a) -> m a

Eliminate a MonadError constraint by providing a continuation should an error be thrown.

This is similar to catchError from Control.Monad.Error.Class, but less confusing.

Eliminating MonadPlus constraints

maybeT :: Monad m => m b -> (a -> m b) -> MaybeT m a -> m b

succeeds :: Monad m => MaybeT m a -> m Bool

Analogous to isJust, but for MaybeT

fails :: Monad m => MaybeT m a -> m Bool

Analogous to isNothing, but for MaybeT